#97 chore: company - DDD 패키지 구색 맞추기
