+++ company-service/src/main/java/com/fn/eureka/client/companyservice/presentation/CompanyController.java
@@ -18,9 +18,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.fn.common.global.util.PageUtils;
-import com.fn.eureka.client.companyservice.application.CompanyService;
-import com.fn.eureka.client.companyservice.presentation.dto.CompanyRequestDto;
-import com.fn.eureka.client.companyservice.presentation.dto.CompanyResponseDto;
+import com.fn.eureka.client.companyservice.domain.service.CompanyService;
+import com.fn.eureka.client.companyservice.presentation.request.CompanyRequestDto;
+import com.fn.eureka.client.companyservice.application.dto.CompanyResponseDto;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
